% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_additional_covariates}
\alias{check_additional_covariates}
\title{Check non-compositional inputs}
\usage{
check_additional_covariates(
  additional_covariates,
  n,
  w_additional_covariates,
  p_x
)
}
\arguments{
\item{additional_covariates}{new data matrix
(see \code{additional_covariates} from \code{\link{trac}})}

\item{n}{number of observations}

\item{w_additional_covariates}{weights for the estimation of the coefficients}

\item{p_x}{vector with number of additional non-compositional covariates}
}
\value{
errors if the requirements are not met
}
\description{
Check if the additional non-compositional inputs have NAs and have the same
number of observations as the compositional inputs
}
\keyword{internal}
