% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_method}
\alias{check_method}
\title{Check method input and other hyperparameter}
\usage{
check_method(method, y, rho = 0)
}
\arguments{
\item{method}{The method (see \code{method}
from \code{\link{trac}})}

\item{y}{The outcome (see \code{y}
from \code{\link{trac}})}

\item{rho}{The hyperparameter for huberized loss for classification
(see \code{rho}from \code{\link{trac}})}
}
\value{
list with vector indicating which method is used and the outcome
}
\description{
Check if the method input and other hyperparameter for classification
are correctly specified
}
\keyword{internal}
