% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classo_fitting.R
\name{classo_fitting}
\alias{classo_fitting}
\title{Perform sparse log-contrast regression}
\usage{
classo_fitting(
  X,
  y,
  C = NULL,
  fraclist = NULL,
  nlam = 20,
  min_frac = 1e-04,
  method = c("regr", "classif", "classif_huber"),
  w = NULL,
  intercept = TRUE,
  rho = 0
)
}
\arguments{
\item{X}{n by p matrix containing e.g. log-ratios}

\item{y}{n vector (response)}

\item{C}{m by p matrix. Default is a row vector of zeros.}

\item{fraclist}{(optional) vector of tuning parameter multipliers.
Should be in (0, 1].}

\item{nlam}{number of tuning parameters (ignored if fraclist non-NULL)}

\item{min_frac}{smallest value of tuning parameter multiplier (ignored if
fraclist non-NULL)}

\item{method}{string which estimation method to use should be in
("regr", "classif", "classif_huber")}

\item{w}{of positive weights of
length ncol(X) (default: all equal to 1).}

\item{intercept}{only works for classification! Should the intercept be
fitted. Default is TRUE, set to FALSE if the intercept should not be
included}

\item{rho}{value for huberized classification loss.
Default = 0.0.}
}
\description{
Solves the constrained lasso problem using the CLASSO module
in Python.  The optimization problem are
}
\details{
Regression:
minimize_{beta, beta0} 1/(2n) || y - beta0 1_n - Zagg_clr beta ||^2
+ lamda_max * frac || beta ||\emph{1
subject to C beta = 0
Classification:
minimize}{beta, beta0} max(1 - y_i(beta0 + Z_clr_i * beta), 0)^2 +
lambda_max * frac || W * beta ||_1
subject to C beta = 0

Default is C = 1_p^T, but C can be a general matrix.

Observe that the tuning parameter is specified through "frac", the fraction
of lamda_max (which is the smallest value for which beta is nonzero).
}
