% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cps_moisture}
\alias{cps_moisture}
\title{Central Park soil moisture data}
\format{
A named list:
\describe{
\item{y}{Vector of n = 580 soil moisture values}
\item{x}{Matrix of soil 16S rRNA amplicon data, with n = 580 rows
corresponding to soil samples and p = 1492 columns corresponding to
families}
\item{tree}{Taxonomic tree of class \code{phylo}}
\item{tax}{A data frame containing the taxonomic information for each family}
\item{A}{A binary matrix encoding the tree structure with p = 1492 rows,
corresponding to leaves, and 1634 columns, corresponding to all non-root
nodes in the tree.}
\item{covariates}{A data frame with the n = 580 soil samples as rows and
five numeric soil properties from the original sample data: \code{pH}, total
carbon (\code{C}), total nitrogen (\code{N}), their ratio (\code{CN}) and \code{CO2_C}. \code{C},
\code{N} and \code{CN} are missing for 29 samples.} }
}
\usage{
cps_moisture
}
\description{
This is the Central Park Soil data used in \href{https://www.biorxiv.org/content/10.1101/2020.09.01.277632v1.full}{Tree-Aggregated Predictive Modeling of Microbiome Data}, with
soil moisture as the response, aggregated to the family level. The soil
samples were collected in New York City's Central Park by \href{https://doi.org/10.1098/rspb.2014.1988}{Ramirez et al. (2014)}.  The counts were summed to
the family level with \code{\link[=aggregate_to_level]{aggregate_to_level()}}, as in the trac paper.
}
\keyword{datasets}
