% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_log_contrast.R
\name{cv_sparse_log_contrast}
\alias{cv_sparse_log_contrast}
\title{Perform cross validation for tuning parameter selection for sparse log contrast}
\usage{
cv_sparse_log_contrast(
  fit,
  Z,
  y,
  folds = NULL,
  nfolds = 5,
  summary_function = stats::median
)
}
\arguments{
\item{fit}{output of \code{\link{sparse_log_contrast}} function.}

\item{Z, y}{same arguments as passed to \code{\link{sparse_log_contrast}}. C
that is used will be taken from fit object.}

\item{folds}{a partition of \code{1:nrow(Z)}.}

\item{nfolds}{number of folds for cross-validation}

\item{summary_function}{how to combine the errors calculated on each
observation within a fold (e.g. mean or median)}
}
\description{
This function is to be called after calling \code{\link{sparse_log_contrast}}.
It performs \code{nfold}-fold cross validation.
}
