% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_categorical_variables}
\alias{get_categorical_variables}
\title{Check if additional variables are categorical}
\usage{
get_categorical_variables(X)
}
\arguments{
\item{X}{see \code{additional_covariates} from \code{\link{trac}}}
}
\value{
list with vector indicating if categorical or not, number of
categorical variables
}
\description{
Check if the additional non-compositional covariates are categorical or not
based on the type of the column. If the column is a binary factor then it is
assumed to be a categorical variable. Useful to transform the input.
}
\keyword{internal}
