% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_probability_cv}
\alias{get_probability_cv}
\title{5-fold cross-validation to find the optimal parameters A and B
for probabilistic output. Obtain the decision values and pass to the platt
algorithm for each lambda}
\usage{
get_probability_cv(
  Z,
  additional_covariates,
  A,
  y,
  method,
  w,
  w_additional_covariates,
  fraclist,
  nfolds = 5,
  eps,
  n_lambda
)
}
\arguments{
\item{Z}{same arguments as passed to \code{\link{trac}}}

\item{additional_covariates}{same arguments as passed to \code{\link{trac}}}

\item{A}{same arguments as passed to \code{\link{trac}}}

\item{y}{same arguments as passed to \code{\link{trac}}}

\item{method}{same arguments as passed to \code{\link{trac}}}

\item{w}{same arguments as passed to \code{\link{trac}}}

\item{w_additional_covariates}{same arguments as passed to \code{\link{trac}}}

\item{fraclist}{same arguments as passed to \code{\link{trac}}}

\item{nfolds}{number of folds}

\item{eps}{convergence criterion}

\item{n_lambda}{number of lambdas}
}
\value{
matrix with hyperparameter A and B as columns
}
\description{
5-fold cross-validation to find the optimal parameters A and B
for probabilistic output. Obtain the decision values and pass to the platt
algorithm for each lambda
}
\keyword{internal}
