% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_probability_platt}
\alias{get_probability_platt}
\title{Calculate parameters A and B based on Platt}
\usage{
get_probability_platt(decision_values, label, eps)
}
\arguments{
\item{decision_values}{output score of trac for classification
(see \code{y} from \code{\link{trac}})}

\item{label}{labels see \code{y} from \code{\link{trac}}}

\item{eps}{threshold for convergence}
}
\value{
A and B
}
\description{
This algorithm is based on the pseudo-code of
}
\keyword{internal}
