% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_trac.R
\name{make_folds_stratified}
\alias{make_folds_stratified}
\title{This function creates stratified folds for cross validation for unbalanced
data. The code is adopted from ggb:::make_folds}
\usage{
make_folds_stratified(n, nfolds, y)
}
\arguments{
\item{n}{number of observations}

\item{nfolds}{number of folds}

\item{y}{variable with group assignment.}
}
\description{
This function creates stratified folds for cross validation for unbalanced
data. The code is adopted from ggb:::make_folds
}
