% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalization_additional_covariates}
\alias{normalization_additional_covariates}
\title{Functions transform non-compositional input}
\usage{
normalization_additional_covariates(additional_covariates, p_x, intercept)
}
\arguments{
\item{additional_covariates}{new data matrix
(see \code{additional_covariates} from \code{\link{trac}})}

\item{p_x}{number of additional covariates}

\item{intercept}{Boolean if normalization should be applied}
}
\value{
list with vector indicating if categorical or not, number of
categorical variables, vector of means and l2 norm and normalized
variables
}
\description{
normalize metadata. (x-mean(x)) / norm(x) for numerical variables and
0/1 for categorical variables provided as a factor
}
\keyword{internal}
