% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cv_trac}
\alias{plot_cv_trac}
\title{Make a plot of the output of cv_trac}
\usage{
plot_cv_trac(cvfit_trac, iw = NULL, superimpose = TRUE)
}
\arguments{
\item{cvfit_trac}{an object as returned by \code{\link{cv_trac}}.}

\item{iw}{vector of indices specifying which weight sequence solutions to
include}

\item{superimpose}{whether to superimpose}
}
\description{
Make a plot of the output of cv_trac
}
