% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trac_path}
\alias{plot_trac_path}
\title{Plot trac coefficient path}
\usage{
plot_trac_path(fit, iw = 1, coef = c("alpha", "beta", "gamma"))
}
\arguments{
\item{fit}{an object as returned by \code{\link{trac}}.}

\item{iw}{index specifying which weight sequence solution to plot}

\item{coef}{which type of parameter to show in the plot}
}
\description{
Plot trac coefficient path
}
