% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_stage.R
\name{predict_second_stage}
\alias{predict_second_stage}
\title{Make predictions based on a second_stage fit}
\usage{
predict_second_stage(
  new_Z,
  new_additional_covariates = NULL,
  fit,
  output = c("raw", "probability", "class")
)
}
\arguments{
\item{new_Z}{a new data matrix (see \code{Z} from \code{\link{second_stage}})}

\item{new_additional_covariates}{a new data matrix
(see \code{additional_covariates} from \code{\link{second_stage}})}

\item{fit}{output of the function \code{\link{second_stage}}}

\item{output}{string  either "raw", "probability" or "class" only relevant
classification tasks and glmnet in second stage}
}
\value{
a vector of \code{nrow(new_Z) + nrow(new_additional_covariates)}
predictions.
}
\description{
Make predictions based on a second_stage fit
}
