% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{probability_transform}
\alias{probability_transform}
\title{Transform the score output to probabilistic output}
\usage{
probability_transform(yhat, A, B)
}
\arguments{
\item{yhat}{score based on the output of trac
(see \code{y} from \code{\link{trac}})}

\item{A}{Hyperparameter for probabilistic transformation}

\item{B}{Hyperparameter for probabilistic transformation}
}
\value{
Probabilistic output
}
\description{
following the idea of Chapter 3.2 of
}
\keyword{internal}
