% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit_log_contrast.R
\name{refit_sparse_log_contrast}
\alias{refit_sparse_log_contrast}
\title{Refit subject to sparsity constraints}
\usage{
refit_sparse_log_contrast(fit, Z, y, tol = 1e-05)
}
\arguments{
\item{fit}{output of trac}

\item{Z, y}{same arguments as passed to \code{\link{sparse_log_contrast}}}

\item{tol}{tolerance for deciding whether a beta value is zero}
}
\description{
Given output of \code{\link{sparse_log_contrast}}, solves the least squares problem with
compositional constraint on the features selected by sparse_log_contrast.
}
\details{
minimize_{beta, beta0} 1/(2n) || y - beta0 1_n - Z beta ||^2
subject to beta_{nonselected} = 0, 1_p^T beta = 0
}
