% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit_log_contrast.R
\name{refit_sparse_log_contrast_classif}
\alias{refit_sparse_log_contrast_classif}
\title{Refit log-contrast for classification to sparsity constraint}
\usage{
refit_sparse_log_contrast_classif(
  fit,
  i_selected = NULL,
  Z,
  y,
  additional_covariates = NULL,
  tol = 1e-05,
  component_selected = NULL
)
}
\arguments{
\item{fit}{output of sparse_log_contrast}

\item{i_selected}{indicator which lambda is selected based on for example the
cross-validation procedure}

\item{Z, y, additional_covariates}{same arguments as passed to
\code{\link{sparse_log_contrast}}}

\item{tol}{tolerance for deciding whether a beta value is zero}

\item{component_selected}{vector with indices which component to include}
}
\description{
Given output of \code{\link{sparse_log_contrast}}, solves the classification
problem with compositional constraint on the features selected by
sparse_log_contrast. In contrast to \code{\link{refit_sparse_log_contrast}}
this function only does the refit for one model specified by i_selected
or component_selected.
This i_selected usually comes from the i1se from the cross-validation output.
}
