% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit_log_contrast.R
\name{refit_sparse_log_contrast_classo}
\alias{refit_sparse_log_contrast_classo}
\title{Refit log-contrast regression/classification to sparsity constraint}
\usage{
refit_sparse_log_contrast_classo(
  fit,
  Z,
  y,
  additional_covariates = NULL,
  tol = 1e-05
)
}
\arguments{
\item{fit}{output of sparse_log_contrast}

\item{Z, y, additional_covariates}{same arguments as passed to
\code{\link{sparse_log_contrast}}}

\item{tol}{tolerance for deciding whether a beta value is zero}
}
\description{
Given output of \code{\link{sparse_log_contrast}}, solves the regression
problem with compositional constraint on the features selected by
sparse_log_contrast. In contrast to \code{\link{refit_sparse_log_contrast}}
this function refit the models based on c-lasso. Be aware that this procedure
does not guard against sign switches meaning that the signs of the
coefficients can change during the refitting procedure.
}
