% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit.R
\name{refit_trac}
\alias{refit_trac}
\title{Refit subject to sparsity constraints}
\usage{
refit_trac(fit, Z, y, A, tol = 1e-05)
}
\arguments{
\item{fit}{output of trac}

\item{Z, y, A}{same arguments as passed to \code{\link{trac}}}

\item{tol}{tolerance for deciding whether a gamma value is zero}
}
\description{
Given output of \code{\link{trac}}, solves the least squares problem with
compositional constraint on the features selected by trac.
}
\details{
minimize_{beta, beta0, gamma} 1/(2n) || y - beta0 1_n - Z beta ||^2
subject to gamma_{trac nonselected} = 0, beta = A gamma, 1_p^T beta = 0
}
