% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescale_betas}
\alias{rescale_betas}
\title{Re-scale betas of normalized variables}
\usage{
rescale_betas(beta, p_x, p, n_numeric, categorical, xs, xm)
}
\arguments{
\item{beta}{estimated coefficients}

\item{p_x}{count of additional non-compositional covariates}

\item{p}{count of compositional covariates}

\item{n_numeric}{count of numerical non-compositional covariates}

\item{categorical}{boolean vector indicating if the additional
non-compositional covariates are categorical or not (see
\code{get_categorical_variables})}

\item{xs}{vector l2 norm (see\code{normalization_additional_covariates})}

\item{xm}{vector of means (see\code{normalization_additional_covariates})}
}
\value{
matrix of re-scaled betas
}
\description{
Re-scales betas of the normalized covariates to the original scale for
non-zero betas
}
\keyword{internal}
