% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getting_A.R
\name{tax_table_to_phylo}
\alias{tax_table_to_phylo}
\title{Convert a tax table to a phylo object}
\usage{
tax_table_to_phylo(x, data = parent.frame(), collapse = TRUE, ...)
}
\arguments{
\item{x}{a right-side formula describing the taxonomic relationship:
    \code{~C1/C2/.../Cn}.}

\item{data}{the data.frame where to look for the variables (default
    to user's workspace).}

\item{collapse}{a logical value specifying whether to collapse single
    nodes in the returned tree (see details).}

\item{...}{further arguments to be passed from other methods.}
}
\description{
This is mostly the function \code{\link[ape]{as.phylo.formula}} from the ape package
but modified so that interior nodes get labeled.
Note: assumes that there are NOT single quotes around the labels.
}
\author{
Julien Dutheil \email{dutheil@evolbio.mpg.de} and Eric Marcon, modified by Jacob Bien
}
