% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_categorical_variables}
\alias{transform_categorical_variables}
\title{Transform non-compositional categorical input}
\usage{
transform_categorical_variables(X, categorical)
}
\arguments{
\item{X}{see \code{additional_covariates} from \code{\link{trac}}}

\item{categorical}{vector indicating if categorical or not}
}
\value{
data frame with transformed categorical variables from factor to
numerical
}
\description{
Transform factors to numerical values of 0 and 1
}
\keyword{internal}
