% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_trac.R
\name{cv_trac}
\alias{cv_trac}
\title{Perform cross validation for tuning parameter selection}
\usage{
cv_trac(fit, Z, y, A, folds = NULL, nfolds = 5)
}
\arguments{
\item{fit}{output of \code{\link{trac}} function.}

\item{Z, y, A}{same arguments as passed to \code{\link{trac}}}

\item{folds}{a partition of \code{1:nrow(X)}.}

\item{nfolds}{number of folds for cross-validation}
}
\description{
This function is to be called after calling \code{\link{trac}}.  It performs
\code{nfold}-fold cross validation.
}
