% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_log_contrast.R
\name{predict_sparse_log_contrast}
\alias{predict_sparse_log_contrast}
\title{Make predictions based on a sparse log contrast fit}
\usage{
predict_sparse_log_contrast(
  fit,
  new_Z,
  new_additional_covariates = NULL,
  output = c("raw", "probability", "class")
)
}
\arguments{
\item{fit}{output of the function \code{\link{sparse_log_contrast}}}

\item{new_Z}{a new data matrix (see \code{Z} from
\code{\link{sparse_log_contrast}})}

\item{new_additional_covariates}{a new data matrix
(see \code{additional_covariates}
from \code{\link{sparse_log_contrast}})}

\item{output}{string  either "raw", "probability" or "class" only relevant
classification tasks}
}
\value{
a vector of \code{nrow(new_Z)} predictions.
}
\description{
Make predictions based on a sparse log contrast fit
}
