% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_stage.R
\name{second_stage}
\alias{second_stage}
\title{Two stage regression}
\usage{
second_stage(
  Z,
  A = NULL,
  y,
  additional_covariates = NULL,
  betas,
  topk = NULL,
  nfolds = 5,
  method = c("regr", "classif"),
  criterion = c("1se", "min"),
  alpha = 0
)
}
\arguments{
\item{Z}{n by p matrix containing log(X) (see \code{Z} from
\code{\link{trac}})}

\item{A}{p by (t_size-1) binary matrix giving tree structure (t_size is the
total number of nodes and the -1 is because we do not include the root).
Only needed for trac based models. (see \code{A} from \code{\link{trac}}).
If A = NULL, a sparse-log contrast model is assumed.}

\item{y}{n vector (response) (see \code{y} from \code{\link{trac}})}

\item{additional_covariates}{n by p' matrix containing additional covariates,
(see \code{additional_covariates} from \code{\link{trac}}).
Non-compositional components are currently not penalized by the lasso}

\item{betas}{pre-screened coefficients
(see output \code{gamma} from \code{\link{trac}} or
output \code{beta} from \code{\link{sparse_log_contrast}})}

\item{topk}{maximum number of pre-screened coefficients to consider.
Default NULL.}

\item{nfolds}{number of folds}

\item{method}{string which estimation method to use should be "regr" or
"classif"}

\item{criterion}{which criterion should be used to select the coefficients
"1se" or "min"}

\item{alpha}{nudge the model to select on a higher or lower level of the
tree. Only relevant for trac based models.}
}
\value{
list with: log_ratios: betas for log ratios; index: dataframe with
index of the pre selected coefficients and the log ratio name; A: taxonomic
tree information; method: regression or classification; cv_glmnet:
output of glmnet, useful for prediction; criterion: which criterion to
be used to select lambda based on cv (cross validation)
}
\description{
Perform a two stage fitting procedure similar to the one proposed by
Bates, S., & Tibshirani, R. (2019). Log‐ratio lasso:
Scalable, sparse estimation for log‐ratio models. Biometrics, 75(2), 613-624.
This function uses the selected components by trac or log-ratio and
fits a sparse model based on all possible log-ratios.
}
\details{
\enumerate{
\item Fit trac or sparse-log contrast and extract the selected components.
\item Fit a second-stage based on the selected components.
}
}
