% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_contrast.R
\name{sparse_log_contrast}
\alias{sparse_log_contrast}
\title{Perform sparse log-contrast regression}
\usage{
sparse_log_contrast(
  Z,
  y,
  C = NULL,
  fraclist = NULL,
  nlam = 20,
  min_frac = 1e-04
)
}
\arguments{
\item{Z}{n by p matrix containing log(X)}

\item{y}{n vector (response)}

\item{C}{m by p matrix. Default is a row vector of ones.}

\item{fraclist}{(optional) vector of tuning parameter multipliers.  Should be in (0, 1].}

\item{nlam}{number of tuning parameters (ignored if fraclist non-NULL)}

\item{min_frac}{smallest value of tuning parameter multiplier (ignored if
fraclist non-NULL)}
}
\description{
Solves the constrained lasso problem using the CLASSO module
in Python.  The optimization problem is
}
\details{
minimize_{beta, beta0} 1/(2n) || y - beta0 1_n - Zagg_clr beta ||^2
+ lamda_max * frac || beta ||_1
subject to C beta = 0

Default is C = 1_p^T, but C can be a general matrix.

Observe that the tuning parameter is specified through "frac", the fraction
of lamda_max (which is the smallest value for which beta is nonzero).
}
